+++ src/main/kotlin/katartal/model/method/instruction/LazyInstructionBuilder.kt
@@ -4,18 +4,18 @@ import katartal.model.ByteCode
 
 class LazyInstructionBuilder internal constructor(
     code: ByteCode,
-    private val reserve: Int,
+    private val reserve: UShort,
     private val evaluator: LazyInstructionBuilder.() -> Unit
 ) :
     InstructionBuilder(code) {
 
-    override val size: Int
-        get() = 1 + reserve
+    override val size: UShort
+        get() = (1u + reserve).toUShort()
 
     override fun flush() {
         this.evaluator()
 
-        if (operands.size != reserve) {
+        if (operands.size.toUShort() != reserve) {
             throw IllegalStateException(
                 "Evaluated operand size (${operands.size}) differs from expected ($reserve). " +
                         "Branch offset instructions might drift."
